refactor(feishu-doc): split doc tools into modular structure
